Export a domain vocabulary; demote wire formats to library internals
The v1 message shapes move to wire.ts, which the public entry point no
longer exports — they are the protocol's encoding, not the contract.
What the library exports instead is the domain vocabulary the app
should speak: EXPOSURE_MODES, FOCUS_MODES, FOCUS_CONTROLS,
ZOOM_CONTROLS, and WHITE_BALANCE_MODES, named after the backend's
enums in icstypes.py. Their values still equal the wire encodings —
sentinel/boolean-string normalization of incoming telemetry is a
separate, later step — but consumers now treat them as opaque
constants.

Two camera-handle refinements follow from the vocabulary: zoom() takes
its speed as a real parameter instead of a "TC:3" suffix baked in by
callers, and triggerOnePushWhiteBalance() expresses what was previously
a magic setWhiteBalance value (the backend models it as an action, not
a mode).
