+++ react-frontend/src/lib/imaging-client/index.ts
@@ -2,7 +2,10 @@
  * Imaging System client library.
  *
  * This is the only module the application may import from; everything else
- * in this directory is an implementation detail of the wire protocol.
+ * in this directory is an implementation detail of the wire protocol. In
+ * particular, the v1 message shapes in wire.ts are internal — the durable
+ * contract is the client API (stations, camera handles, typed channels)
+ * plus the domain vocabulary exported here.
  */
 
 export { createImagingClient, getSharedImagingClient } from "./client";
@@ -16,6 +19,21 @@ export type {
 } from "./client";
 
 export {
+  EXPOSURE_MODES,
+  FOCUS_CONTROLS,
+  FOCUS_MODES,
+  WHITE_BALANCE_MODES,
+  ZOOM_CONTROLS,
+} from "./domain";
+export type {
+  ExposureMode,
+  FocusControl,
+  FocusMode,
+  WhiteBalanceMode,
+  ZoomControl,
+} from "./domain";
+
+export {
   coerceObserverSide,
   getObserverInfo,
   normalizeObserverSide,
@@ -26,31 +44,10 @@ export {
 export type { ObserverSide, ObserverSideInput, ObserverInfo } from "./protocol";
 
 export type {
-  CamHeartbeat,
-  CameraArrayEntry,
-  CameraCommandAction,
-  CameraCommandBody,
-  CameraCommandPayload,
-  CameraSettingsCurrent,
-  CameraSettingsPayload,
-  CommandReceipt,
   ConnectionStatus,
   ConnectionStatusEvent,
-  ExposureMode,
-  FocusControl,
-  FocusMode,
-  NavHeartbeat,
-  RecorderHeartbeat,
-  RecorderHeartbeatObserver,
-  RecorderHeartbeatPilot,
-  ReceiptStatus,
-  SensorHeartbeat,
   SentCommand,
-  SystemMessage,
   Unsubscribe,
-  User,
-  WhiteBalanceMode,
   WsEndpoint,
   WsEndpoints,
-  ZoomControl,
 } from "./types";
